[frontend] move Graphs components into dedicated folder
