+++ opencti-platform/opencti-front/src/private/components/analyses/groupings/GroupingKnowledgeGraph.jsx
@@ -37,13 +37,13 @@ import { groupingMutationFieldPatch } from './GroupingEditionOverview';
 import { groupingKnowledgeGraphMutationRelationDeleteMutation, groupingKnowledgeGraphtMutationRelationAddMutation } from './GroupingKnowledgeGraphQuery';
 import ContainerHeader from '../../common/containers/ContainerHeader';
 import GroupingPopover from './GroupingPopover';
-import EntitiesDetailsRightsBar from '../../../../utils/graph/EntitiesDetailsRightBar';
-import LassoSelection from '../../../../utils/graph/LassoSelection';
+import EntitiesDetailsRightsBar from '../../../../utils/graph/components/EntitiesDetailsRightBar';
+import LassoSelection from '../../../../utils/graph/components/LassoSelection';
 import { hexToRGB } from '../../../../utils/Colors';
 import { UserContext } from '../../../../utils/hooks/useAuth';
 import investigationAddFromContainer from '../../../../utils/InvestigationUtils';
 import { isNotEmptyField } from '../../../../utils/utils';
-import RelationSelection from '../../../../utils/graph/RelationSelection';
+import RelationSelection from '../../../../utils/graph/components/RelationSelection';
 import withRouter from '../../../../utils/compat_router/withRouter';
 import { containerTypes } from '../../../../utils/hooks/useAttributes';
