+++ opencti-platform/opencti-front/src/private/components/cases/case_rfts/CaseRftKnowledgeCorrelation.jsx
@@ -24,12 +24,12 @@ import {
   nodePaint,
   nodeThreePaint,
 } from '../../../../utils/Graph';
-import LassoSelection from '../../../../utils/graph/LassoSelection';
+import LassoSelection from '../../../../utils/graph/components/LassoSelection';
 import { buildViewParamsFromUrlAndStorage, saveViewParameters } from '../../../../utils/ListParameters';
 import { caseRftMutationFieldPatch } from './CaseRftEditionOverview';
 import { UserContext } from '../../../../utils/hooks/useAuth';
 import CaseRftKnowledgeGraphBar from './CaseRftKnowledgeGraphBar';
-import EntitiesDetailsRightsBar from '../../../../utils/graph/EntitiesDetailsRightBar';
+import EntitiesDetailsRightsBar from '../../../../utils/graph/components/EntitiesDetailsRightBar';
 import { isNotEmptyField } from '../../../../utils/utils';
 import { getMainRepresentative, getSecondaryRepresentative } from '../../../../utils/defaultRepresentatives';
