+++ opencti-platform/opencti-front/src/utils/graph/components/LassoSelection.tsx
@@ -3,8 +3,8 @@ import { SimplePaletteColorOptions } from '@mui/material';
 import { useTheme } from '@mui/styles';
 import makeStyles from '@mui/styles/makeStyles';
 import { ForceGraphMethods } from 'react-force-graph-2d';
-import type { Theme } from '../../components/Theme';
-import { pointInPolygon } from '../Graph';
+import type { Theme } from '../../../components/Theme';
+import { pointInPolygon } from '../../Graph';
 
 // Deprecated - https://mui.com/system/styles/basics/
 // Do not use it for new code.
