+++ opencti-platform/opencti-front/src/utils/graph/components/RelationSelection.tsx
@@ -3,7 +3,7 @@ import { useTheme } from '@mui/styles';
 import makeStyles from '@mui/styles/makeStyles';
 import React, { FunctionComponent, MutableRefObject, useCallback, useEffect, useRef } from 'react';
 import { ForceGraphMethods } from 'react-force-graph-2d';
-import type { Theme } from '../../components/Theme';
+import type { Theme } from '../../../components/Theme';
 
 // Deprecated - https://mui.com/system/styles/basics/
 // Do not use it for new code.
