+++ opencti-platform/opencti-front/src/utils/graph/components/StixMetaObjectDetails.tsx
@@ -3,17 +3,17 @@ import makeStyles from '@mui/styles/makeStyles';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
 import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
-import { useFormatter } from '../../components/i18n';
-import useQueryLoading from '../hooks/useQueryLoading';
-import Loader, { LoaderVariant } from '../../components/Loader';
+import { useFormatter } from '../../../components/i18n';
+import useQueryLoading from '../../hooks/useQueryLoading';
+import Loader, { LoaderVariant } from '../../../components/Loader';
 import type { SelectedEntity } from './EntitiesDetailsRightBar';
-import ErrorNotFound from '../../components/ErrorNotFound';
-import { getMainRepresentative } from '../defaultRepresentatives';
-import { truncate } from '../String';
-import ItemCreators from '../../components/ItemCreators';
-import { StixMetaObjectDetailsQuery } from './__generated__/StixMetaObjectDetailsQuery.graphql';
-import ItemMarkings from '../../components/ItemMarkings';
-import ItemEntityType from '../../components/ItemEntityType';
+import ErrorNotFound from '../../../components/ErrorNotFound';
+import { getMainRepresentative } from '../../defaultRepresentatives';
+import { truncate } from '../../String';
+import ItemCreators from '../../../components/ItemCreators';
+import { StixMetaObjectDetailsQuery } from '../__generated__/StixMetaObjectDetailsQuery.graphql';
+import ItemMarkings from '../../../components/ItemMarkings';
+import ItemEntityType from '../../../components/ItemEntityType';
 
 // Deprecated - https://mui.com/system/styles/basics/
 // Do not use it for new code.
